+++ packages/terra-core-docs/src/terra-dev-site/doc/alert/example/AdvisoryExample.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Alert from 'terra-alert';
 
 const AdvisoryExample = () => (
-  <Alert type="advisory">This is an advisory alert</Alert>
+  <Alert type="advisory">This is an advisory notification banner</Alert>
 );
 
 export default AdvisoryExample;