+++ packages/terra-scroll/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added
   * Added styles when the scroll is focussed
+* Changed
+  * Moved the focus style changes from terra-scroll to terra-content-continer
 
 ## 2.34.0 - (February 15, 2023)