+++ packages/terra-scroll/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added
   * Added styles when the scroll is focussed
+* Changed
+  * Reverted the focus style changes added in terra-scroll.
 
 ## 2.34.0 - (February 15, 2023)